+++ react/components/icon/Plus/index.js
@@ -1,6 +1,5 @@
 import React, { PureComponent } from 'react'
 import PropTypes from 'prop-types'
-import config from 'vtex-tachyons/config.json'
 import { calcIconSize, baseClassname } from '../utils'
 
 const iconBase = {
@@ -35,7 +34,7 @@ class Plus extends PureComponent {
 }
 
 Plus.defaultProps = {
-  color: config.colors['serious-black'],
+  color: 'currentColor',
   size: 16,
   block: false,
 }